INTRD-12798 fix quantity of 0,5 is not allowed

+++ opencell-api/src/main/java/org/meveo/api/cpq/CommercialOrderApi.java
@@ -1200,7 +1200,7 @@ final CommercialOrder order = commercialOrderService.findById(orderDto.getId());
 	        handleMissingParameters();
 	    }
 		for (OrderProductDto orderProductDto : orderProductDtos) {  
-		    if(orderProductDto.getQuantity() == null || orderProductDto.getQuantity().intValue() == 0 )
+		    if(orderProductDto.getQuantity() == null || orderProductDto.getQuantity().doubleValue() == 0 )
 		        throw new BusinessApiException("The quantity for product code " + orderProductDto.getProductCode() + " must be great than 0" );
 			OrderProduct orderProduct=populateOrderProduct(orderProductDto,orderOffer,null);  
 			orderProductService.create(orderProduct);
